WP-G1a: split remaining Tauri-mixed audio/database/onboarding files into pure-core + pure-shell
Prepares src/audio/**, src/database/**, src/onboarding.rs, and
src/mcp_config.rs for the upcoming meetily-core crate split: every .rs
file in this area is now either free of Tauri types/attributes or a thin
*_commands.rs shell that only builds Tauri glue (AppHandle/State) and
delegates to the core module.

- events.rs: moved the `impl EventSink for tauri::AppHandle<R>` and
  `shared_sink()` into a new src/tauri_events.rs shell module; updated
  every `crate::events::shared_sink` call site across the crate
  (audio, api, ollama, speaker_diarization, summary, whisper_engine,
  lib.rs) to `crate::tauri_events::shared_sink`.
- audio/ffmpeg.rs, audio/incremental_saver.rs, audio/clip.rs: split the
  #[tauri::command] wrappers out into ffmpeg_commands.rs,
  incremental_saver_commands.rs, clip_commands.rs.
- audio/import.rs, audio/retranscription.rs: moved the AppHandle-taking
  `start_import`/`start_retranscription` shell wrapper and all
  #[tauri::command] fns into new import_commands.rs /
  retranscription_commands.rs; the Tauri-free `start_import_with` /
  `start_retranscription_with` core already existed from WP-C/WP-E and
  is now the only entry point left in the core files.
- audio/recording_preferences.rs: `save_recording_preferences` now takes
  `Option<SqlitePool>` instead of `&AppHandle<R>`; db_pool() and all
  #[tauri::command] fns moved to recording_preferences_commands.rs.
- database/repositories/setting.rs: moved the two UI-config commands
  (api_get_ui_config/api_save_ui_config) into database/commands.rs.
- database/setup.rs: replaced `initialize_database_on_startup(&AppHandle)`
  with a Tauri-free `prepare_database_on_startup() -> StartupOutcome`;
  the AppHandle-driven state/emit glue moved to a new
  `initialize_database_on_startup` helper in lib.rs.
- mcp_config.rs: dropped the unused AppHandle param from
  `get_mcp_server_info`; both commands now live in new
  mcp_config_commands.rs.
- onboarding.rs: load/save/reset/fetch now take `Option<SqlitePool>`
  instead of `&AppHandle<R>`; db_pool(), the legacy-JSON import helper,
  and all #[tauri::command] fns (plus complete_onboarding) moved to new
  onboarding_commands.rs; tray.rs now calls
  onboarding_commands::load_onboarding_status_for_app.
- database/manager.rs, recording_manager.rs, recording_service.rs,
  recording_commands.rs, recovery_commands.rs, state.rs: verified
  already Tauri-free (or, for the *_commands.rs files, already pure
  shell) — no changes needed beyond confirming via grep.
- Updated lib.rs's generate_handler![] paths and mod declarations for
  every moved command; command names/signatures are unchanged.

cargo check -p meetily: zero warnings. cargo test -p meetily --lib:
300 passed, 0 failed (matches baseline), 3 ignored.
