+++ frontend/src-tauri/src/audio/incremental_saver.rs
@@ -326,7 +326,6 @@ pub struct AudioRecoveryStatus {
 /// This is called by the transcript recovery system after a crash: PCM
 /// checkpoints are encoded into `audio.mp4` in one pass; legacy AAC
 /// checkpoints are stream-copy concatenated.
-#[tauri::command]
 pub async fn recover_audio_from_checkpoints(
     meeting_folder: String,
     _sample_rate: u32,
@@ -509,7 +508,6 @@ pub async fn recover_audio_from_checkpoints(
 
 /// Clean up checkpoint files after successful recording or recovery
 /// This command is called by the frontend after successful save to clean up checkpoint files
-#[tauri::command]
 pub async fn cleanup_checkpoints(meeting_folder: String) -> Result<(), String> {
     info!("Cleaning up checkpoints for folder: {}", meeting_folder);
 
@@ -529,7 +527,6 @@ pub async fn cleanup_checkpoints(meeting_folder: String) -> Result<(), String> {
 
 /// Check if a meeting folder has audio checkpoint files
 /// Returns true if .checkpoints/ directory exists and contains .mp4 files
-#[tauri::command]
 pub async fn has_audio_checkpoints(meeting_folder: String) -> Result<bool, String> {
     let folder_path = PathBuf::from(&meeting_folder);
     let checkpoints_dir = folder_path.join(".checkpoints");
