+++ frontend/src-tauri/src/audio/recording_preferences_commands.rs
@@ -1,0 +1,75 @@
+//! Tauri command wrappers for recording preferences. Core logic (loading,
+//! saving, defaults) lives in `audio::recording_preferences`.
+
+use log::{info, warn};
+use tauri::{AppHandle, Manager, Runtime};
+
+use crate::audio::recording_preferences::{
+    self, ensure_recordings_directory, get_default_recordings_folder, RecordingPreferences,
+};
+use crate::state::AppState;
+
+/// Fetch the SQLite pool, if `AppState` has been managed yet. `None` early
+/// in startup — e.g. on a first-launch cold start, before the frontend has
+/// created the database.
+fn db_pool<R: Runtime>(app: &AppHandle<R>) -> Option<sqlx::SqlitePool> {
+    app.try_state::<AppState>()
+        .map(|s| s.db_manager.pool().clone())
+}
+
+#[tauri::command]
+pub async fn get_recording_preferences<R: Runtime>(
+    app: AppHandle<R>,
+) -> Result<RecordingPreferences, String> {
+    recording_preferences::load_recording_preferences(db_pool(&app))
+        .await
+        .map_err(|e| format!("Failed to load recording preferences: {}", e))
+}
+
+#[tauri::command]
+pub async fn set_recording_preferences<R: Runtime>(
+    app: AppHandle<R>,
+    preferences: RecordingPreferences,
+) -> Result<(), String> {
+    recording_preferences::save_recording_preferences(db_pool(&app), &preferences)
+        .await
+        .map_err(|e| format!("Failed to save recording preferences: {}", e))
+}
+
+#[tauri::command]
+pub async fn get_default_recordings_folder_path() -> Result<String, String> {
+    let path = get_default_recordings_folder();
+    Ok(path.to_string_lossy().to_string())
+}
+
+#[tauri::command]
+pub async fn open_recordings_folder<R: Runtime>(app: AppHandle<R>) -> Result<(), String> {
+    let preferences = recording_preferences::load_recording_preferences(db_pool(&app))
+        .await
+        .map_err(|e| format!("Failed to load preferences: {}", e))?;
+
+    // Ensure directory exists before trying to open it
+    ensure_recordings_directory(&preferences.save_folder)
+        .map_err(|e| format!("Failed to create directory: {}", e))?;
+
+    let folder_path = preferences.save_folder.to_string_lossy().to_string();
+
+    std::process::Command::new("xdg-open")
+        .arg(&folder_path)
+        .spawn()
+        .map_err(|e| format!("Failed to open folder: {}", e))?;
+
+    info!("Opened recordings folder: {}", folder_path);
+    Ok(())
+}
+
+#[tauri::command]
+pub async fn select_recording_folder<R: Runtime>(
+    _app: AppHandle<R>,
+) -> Result<Option<String>, String> {
+    // Use Tauri's dialog to select folder
+    // For now, return None - this would need to be implemented with tauri-plugin-dialog
+    // when it's available in the Cargo.toml
+    warn!("Folder selection not yet implemented - using dialog plugin");
+    Ok(None)
+}
