+++ frontend/src-tauri/src/database/commands.rs
@@ -1,7 +1,8 @@
 use log::{error, info};
-use tauri::{AppHandle, Manager};
+use tauri::{AppHandle, Manager, State};
 
 use super::manager::DatabaseManager;
+use super::repositories::setting::{SettingsRepository, KEY_UI_CONFIG};
 use crate::events::EventSinkExt;
 use crate::state::AppState;
 
@@ -93,4 +94,35 @@ pub async fn open_database_folder(_app: AppHandle) -> Result<(), String> {
 
     info!("Opened database folder: {}", folder_path);
     Ok(())
+}
+
+// ===== FRONTEND UI CONFIG COMMANDS =====
+//
+// Thin Tauri commands backing ConfigContext's previously-localStorage-only
+// preferences (primary language, confidence indicator toggle, auto-summary
+// toggle, per-provider model cache, ...). The shape is owned entirely by the
+// frontend — the backend just persists whatever JSON blob it's handed.
+
+/// Gets the saved frontend UI config blob, or `None` if nothing has been saved yet.
+#[tauri::command]
+pub async fn api_get_ui_config(
+    state: State<'_, AppState>,
+) -> Result<Option<serde_json::Value>, String> {
+    let pool = state.db_manager.pool();
+    SettingsRepository::get_setting_json(pool, KEY_UI_CONFIG)
+        .await
+        .map(|opt| opt.and_then(|json| serde_json::from_str(&json).ok()))
+        .map_err(|e| format!("Failed to load UI config: {}", e))
+}
+
+/// Saves the frontend UI config blob (full replace).
+#[tauri::command]
+pub async fn api_save_ui_config(
+    state: State<'_, AppState>,
+    config: serde_json::Value,
+) -> Result<(), String> {
+    let pool = state.db_manager.pool();
+    SettingsRepository::set_setting(pool, KEY_UI_CONFIG, &config)
+        .await
+        .map_err(|e| format!("Failed to save UI config: {}", e))
 }
